--- a/executor/benchmark_test.go
+++ b/executor/benchmark_test.go
@@ -530,6 +530,7 @@ func BenchmarkWindowFunctions(b *testing.B) {
 
 type hashJoinTestCase struct {
 	rows        int
+	cols        []*types.FieldType
 	concurrency int
 	ctx         sessionctx.Context
 	keyIdx      []int
@@ -537,24 +538,27 @@ type hashJoinTestCase struct {
 }
 
 func (tc hashJoinTestCase) columns() []*expression.Column {
-	return []*expression.Column{
-		{Index: 0, RetType: types.NewFieldType(mysql.TypeLonglong)},
-		{Index: 1, RetType: types.NewFieldType(mysql.TypeVarString)},
+	ret := make([]*expression.Column, 0)
+	for i, t := range tc.cols {
+		column := &expression.Column{Index: i, RetType: t}
+		ret = append(ret, column)
 	}
+	return ret
 }
 
 func (tc hashJoinTestCase) String() string {
-	return fmt.Sprintf("(rows:%v, concurency:%v, joinKeyIdx: %v, disk:%v)",
-		tc.rows, tc.concurrency, tc.keyIdx, tc.disk)
+	return fmt.Sprintf("(rows:%v, cols:%v, concurency:%v, joinKeyIdx: %v, disk:%v)",
+		tc.rows, tc.cols, tc.concurrency, tc.keyIdx, tc.disk)
 }
 
-func defaultHashJoinTestCase() *hashJoinTestCase {
+func defaultHashJoinTestCase(cols []*types.FieldType) *hashJoinTestCase {
 	ctx := mock.NewContext()
 	ctx.GetSessionVars().InitChunkSize = variable.DefInitChunkSize
 	ctx.GetSessionVars().MaxChunkSize = variable.DefMaxChunkSize
 	ctx.GetSessionVars().StmtCtx.MemTracker = memory.NewTracker(nil, -1)
 	ctx.GetSessionVars().IndexLookupJoinConcurrency = 4
 	tc := &hashJoinTestCase{rows: 100000, concurrency: 4, ctx: ctx, keyIdx: []int{0, 1}}
+	tc.cols = cols
 	return tc
 }
 
@@ -606,6 +610,8 @@ func benchmarkHashJoinExecWithCase(b *testing.B, casTest *hashJoinTestCase) {
 				return int64(row)
 			case mysql.TypeVarString:
 				return rawData
+			case mysql.TypeDouble:
+				return float64(row)
 			default:
 				panic("not implement")
 			}
@@ -651,8 +657,13 @@ func BenchmarkHashJoinExec(b *testing.B) {
 	log.SetLevel(zapcore.ErrorLevel)
 	defer log.SetLevel(lvl)
 
+	cols := []*types.FieldType{
+		types.NewFieldType(mysql.TypeLonglong),
+		types.NewFieldType(mysql.TypeVarString),
+	}
+
 	b.ReportAllocs()
-	cas := defaultHashJoinTestCase()
+	cas := defaultHashJoinTestCase(cols)
 	b.Run(fmt.Sprintf("%v", cas), func(b *testing.B) {
 		benchmarkHashJoinExecWithCase(b, cas)
 	})
@@ -671,6 +682,21 @@ func BenchmarkHashJoinExec(b *testing.B) {
 	cas.keyIdx = []int{0}
 	cas.disk = true
 	cas.rows = 1000
+	b.Run(fmt.Sprintf("%v", cas), func(b *testing.B) {
+		benchmarkHashJoinExecWithCase(b, cas)
+	})
+
+	// Replace the wide string column with double column
+	cols = []*types.FieldType{
+		types.NewFieldType(mysql.TypeLonglong),
+		types.NewFieldType(mysql.TypeDouble),
+	}
+	cas = defaultHashJoinTestCase(cols)
+	b.Run(fmt.Sprintf("%v", cas), func(b *testing.B) {
+		benchmarkHashJoinExecWithCase(b, cas)
+	})
+
+	cas.keyIdx = []int{0}
 	b.Run(fmt.Sprintf("%v", cas), func(b *testing.B) {
 		benchmarkHashJoinExecWithCase(b, cas)
 	})
@@ -732,8 +758,13 @@ func BenchmarkBuildHashTableForList(b *testing.B) {
 	log.SetLevel(zapcore.ErrorLevel)
 	defer log.SetLevel(lvl)
 
+	cols := []*types.FieldType{
+		types.NewFieldType(mysql.TypeLonglong),
+		types.NewFieldType(mysql.TypeVarString),
+	}
+
 	b.ReportAllocs()
-	cas := defaultHashJoinTestCase()
+	cas := defaultHashJoinTestCase(cols)
 	rows := []int{10, 100000}
 	keyIdxs := [][]int{{0, 1}, {0}}
 	disks := []bool{false, true}