<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Fix #12048 

### What is changed and how it works?
```
name                                                                       old time/op  new time/op  delta
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-8   716ms ± 1%   717ms ± 3%    ~     (p=1.000 n=16+20)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0],_disk:false)-8     102ms ± 2%   100ms ± 1%  -1.16%  (p=0.000 n=20+18)
HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0],_disk:true)-8      708ms ±13%   700ms ±11%    ~     (p=0.640 n=20+20)
HashJoinExec/(rows:1000,_concurency:4,_joinKeyIdx:_[0],_disk:true)-8       17.8ms ±20%  17.9ms ±15%    ~     (p=0.968 n=20+20)
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes

 - Has exported function/method change

Side effects

 - Possible performance regression
 - Increased code complexity

Related changes

Release note

 - Write release note for bug-fix or new feature.