@SunRunAway 2019-10-17
Hi, @sduzh
Thank you for your PR, your code looks great.

May you help do a little more profiling?

Use `go test -run=XX -bench="BenchmarkHashJoinExec" -benchmem -count 5  -memprofile memprofile.out -cpuprofile profile.out` to collect `profile.out`

`go tool pprof -http=":8080" profile.out` to analyze through flame graph, or use `go tool pprof profile.out` and `list` command in CLI to analyze the cumulative cost of hash calculation during probing. Maybe you would like to change the function name of `getJoinKeyFromChkRow` for probing to get the cumulative cost during probing in old code.

Also, you may just target this test `HashJoinExec/(rows:100000,_concurency:4,_joinKeyIdx:_[0],_disk:false)-8`, and temporarily disable the others.