@sduzh 2019-10-17
@SunRunAway 
results on this new branch
```
➜  executor git:(vec-hash-probe) ✗ go tool pprof profile.out
Type: cpu
Time: Oct 18, 2019 at 1:27am (CST)
Duration: 23.05s, Total samples = 38.50s (167.05%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) list join2Chunk
Total: 38.50s
ROUTINE ======================== github.com/pingcap/tidb/executor.(*HashJoinExec).join2Chunk in /Users/zhuming/go/src/github.com/sduzh/tidb/executor/join.go
     130ms     10.07s (flat, cum) 26.16% of Total
         .          .    421:		return false, joinResult
         .          .    422:	}
         .          .    423:
         .          .    424:	hCtx.initHash(outerChk.NumRows())
         .          .    425:	for _, i := range hCtx.keyColIdx {
         .       20ms    426:		err = codec.HashChunkSelected(e.rowContainer.sc, hCtx.hashVals, outerChk, hCtx.allTypes[i], i, hCtx.buf, hCtx.hasNull, selected)
         .          .    427:		if err != nil {
         .          .    428:			joinResult.err = err
         .          .    429:			return false, joinResult
         .          .    430:		}
         .          .    431:	}
         .          .    432:
      20ms       20ms    433:	for i := range selected {
      50ms       50ms    434:		if !selected[i] || hCtx.hasNull[i] { // process unmatched outer rows
         .          .    435:			e.joiners[workerID].onMissMatch(false, outerChk.GetRow(i), joinResult.chk)
         .          .    436:		} else { // process matched outer rows
      30ms       50ms    437:			probeKey, probeRow := hCtx.hashVals[i].Sum64(), outerChk.GetRow(i)
      30ms      9.93s    438:			ok, joinResult = e.joinMatchedOuterRow2Chunk(workerID, probeKey, probeRow, hCtx, joinResult)
         .          .    439:			if !ok {
         .          .    440:				return false, joinResult
         .          .    441:			}
         .          .    442:		}
         .          .    443:		if joinResult.chk.IsFull() {
```

results on master branch
```
➜  executor git:(master) ✗ go tool pprof profile.out.old
Type: cpu
Time: Oct 18, 2019 at 1:34am (CST)
Duration: 20.49s, Total samples = 33.69s (164.45%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) list join2Chunk
Total: 33.69s
ROUTINE ======================== github.com/pingcap/tidb/executor.(*HashJoinExec).join2Chunk in /Users/zhuming/go/src/github.com/sduzh/tidb/executor/join.go
      30ms      9.85s (flat, cum) 29.24% of Total
         .          .    412:}
         .          .    413:
         .          .    414:func (e *HashJoinExec) join2Chunk(workerID uint, outerChk *chunk.Chunk, hCtx *hashContext, joinResult *hashjoinWorkerResult,
         .          .    415:	selected []bool) (ok bool, _ *hashjoinWorkerResult) {
         .          .    416:	var err error
      10ms       20ms    417:	selected, err = expression.VectorizedFilter(e.ctx, e.outerFilter, chunk.NewIterator4Chunk(outerChk), selected)
         .          .    418:	if err != nil {
         .          .    419:		joinResult.err = err
         .          .    420:		return false, joinResult
         .          .    421:	}
         .          .    422:	for i := range selected {
      10ms       10ms    423:		if !selected[i] { // process unmatched outer rows
         .          .    424:			e.joiners[workerID].onMissMatch(false, outerChk.GetRow(i), joinResult.chk)
         .          .    425:		} else { // process matched outer rows
      10ms      9.82s    426:			ok, joinResult = e.joinMatchedOuterRow2Chunk(workerID, outerChk.GetRow(i), hCtx, joinResult)
         .          .    427:			if !ok {
         .          .    428:				return false, joinResult
         .          .    429:			}
         .          .    430:		}
         .          .    431:		if joinResult.chk.IsFull() {
(pprof) list getJoinKeyFromChkRow
Total: 33.69s
ROUTINE ======================== github.com/pingcap/tidb/executor.(*hashRowContainer).getJoinKeyFromChkRowX in /Users/zhuming/go/src/github.com/sduzh/tidb/executor/hash_table.go
      30ms      290ms (flat, cum)  0.86% of Total
         .          .    239:}
         .          .    240:
         .          .    241:// getJoinKeyFromChkRow fetches join keys from row and calculate the hash value.
         .          .    242:func (*hashRowContainer) getJoinKeyFromChkRow(sc *stmtctx.StatementContext, row chunk.Row, hCtx *hashContext) (hasNull bool, key uint64, err error) {
         .          .    243:	for _, i := range hCtx.keyColIdx {
         .       10ms    244:		if row.IsNull(i) {
         .          .    245:			return true, 0, nil
         .          .    246:		}
         .          .    247:	}
      10ms       50ms    248:	hCtx.initHash(1)
      10ms      220ms    249:	err = codec.HashChunkRow(sc, hCtx.hashVals[0], row, hCtx.allTypes, hCtx.keyColIdx, hCtx.buf)
      10ms       10ms    250:	return false, hCtx.hashVals[0].Sum64(), err
         .          .    251:}
         .          .    252:
         .          .    253:// Len returns the length of the records in hashRowContainer.
         .          .    254:func (c hashRowContainer) Len() int {
         .          .    255:	return c.hashTable.Len()
(pprof)
```