@sduzh 2019-10-29
Sorry for the late reply. 
As @SunRunAway mentioned above, the original test case must join the wide string columns, which result in a lot of string copies.
So I designed a new test case, as suggested by @SunRunAway, replace the string columns with double columns, and the result is much better:
```
➜  executor git:(vec-hash-probe) ✗ benchstat old.txt new.txt
name                                                                                                        old time/op  new time/op  delta
HashJoinExec/(rows:100000,_cols:[[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0_1],_disk:false)-8         28.7ms ± 1%  23.5ms ± 1%  -18.10%  (p=0.008 n=5+5)
HashJoinExec/(rows:100000,_cols:[bigint(20)_double],_concurency:4,_joinKeyIdx:_[0],_disk:false)-8           26.2ms ± 3%  22.8ms ± 4%  -12.77%  (p=0.008 n=5+5)

``` 
@XuHuaiyu